+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Enabled autolinking for `nex://` URIs.
 - Display FEP-044f quotes.
+- Signal support for RFC9421+Ed25519.
 
 ### Changed